feat: add edit-event from V1 API
